requirements.txt
flask==2.3.3
opencv-python==4.8.0.76
face-recognition==1.3.0
numpy==1.26.0
pytz==2023.3
reportlab==3.6.12
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-auth==2.22.0
